Removed unnecessary casting into short value type.

+++ src/test/java/GymNotebook/model/SetTest.java
@@ -8,15 +8,15 @@ import static org.junit.jupiter.api.Assertions.*;
 public class SetTest {
     @Test
     public void testSetGettersAndSetters() {
-        RepSet set = new RepSet((short) 15, (short) 120);
+        RepSet set = new RepSet( 15, 120);
 
-        assertEquals((short) 15, set.getRepCount());
-        assertEquals((short) 120, set.getWeight());
+        assertEquals(15, set.getRepCount());
+        assertEquals(120, set.getWeight());
     }
 
     @Test
     public void testSetJson() throws JsonProcessingException {
-        RepSet set = new RepSet((short) 1, (short) 60);
+        RepSet set = new RepSet(1, 60);
 
         ObjectMapper mapper = new ObjectMapper();
